+++ docs/versionhistory.rst
@@ -9,6 +9,7 @@ This library adheres to `Semantic Versioning <https://semver.org/>`_.
 
 - Dropped support for Python 3.8
   (#247 <https://github.com/agronholm/cbor2/pull/247>_; PR by @hugovk)
+- Added support for encoding indefinite containers
 
 **5.6.5** (2024-10-09)